backend/requirements.txt
# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.34.3

# Database
sqlalchemy==2.0.41
sqlmodel==0.0.24
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# File Upload & Storage - OPTIMIERT: Nur notwendige Pakete
redis==5.0.1  # Caching für Performance

# Validation & Serialization
pydantic==2.11.5
pydantic-settings==2.2.1

# Environment & Configuration
python-dotenv==1.1.0

# Monitoring & Performance
psutil==6.1.0  # System monitoring

# Development - Nur für Tests
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0

